// src/app/generated/cite.api/model/itemStatus.ts
/*
Copyright 2022 Carnegie Mellon University. All Rights Reserved. 
 Released under a MIT (SEI)-style license. See LICENSE.md in the
// project root for license information.
*/

/**
 * Cite API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type ItemStatus = 'Pending' | 'Active' | 'Cancelled' | 'Complete' | 'Archived';

export const ItemStatus = {
    Pending: 'Pending' as ItemStatus,
    Active: 'Active' as ItemStatus,
    Cancelled: 'Cancelled' as ItemStatus,
    Complete: 'Complete' as ItemStatus,
    Archived: 'Archived' as ItemStatus
};