// src/app/generated/cite.api/model/teamPermissionClaim.ts
/*
 Copyright 2026 Carnegie Mellon University. All Rights Reserved.
 Released under a MIT (SEI)-style license. See LICENSE.md in the
 project root for license information.
*/

/**
 * Cite API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TeamPermission } from './teamPermission';


export interface TeamPermissionClaim { 
    teamId?: string;
    permissions?: Array<TeamPermission> | null;
}